Updated titlebar options and top-level titlebar.
Fixed bugs with SizeFrame. Ensure that the resize does not go past screen/area boundaries. Also ensure that the widget does not jump, by ensuring we have a local and not global rect. We also fixed that if the resize goes below the minimum size values of the widget, we don't drag the widget.

Added top-level titlebar. This has been tested to work on X11 (with numerous workarounds), and has various tricks to set window flags without hiding and re-showing the window.

Fixed bugs with the menu size event. Properly keep the override cursor outside the application boundaries, and handle clicks outside as well. It also limits the resizing to the current MDI area's bounds.

Still need to provide fixes for moving/resizing windows on Wayland.